# api/build/smithy/source/openapi/Api.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "codineers-project-api",
        "version": "2024-06-20"
    },
    "paths": {
        "/projects": {
            "get": {
                "operationId": "ListProjects",
                "responses": {
                    "200": {
                        "description": "ListProjects 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListProjectsResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "CreateProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "CreateProject 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateProjectResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/projects/{projectId}": {
            "delete": {
                "operationId": "DeleteProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "pattern": "^[a-zA-Z0-9]+$"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DeleteProject 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteProjectResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "pattern": "^[a-zA-Z0-9]+$"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetProject 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetProjectResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "UpdateProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequestContent"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "pattern": "^[a-zA-Z0-9]+$"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UpdateProject 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateProjectResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateProjectRequestContent": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9]+$"
                    },
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                },
                "required": [
                    "client",
                    "department",
                    "endDate",
                    "projectId",
                    "startDate"
                ]
            },
            "CreateProjectResponseContent": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9]+$"
                    },
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                },
                "required": [
                    "projectId"
                ]
            },
            "DeleteProjectResponseContent": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "GetProjectResponseContent": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9]+$"
                    },
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                },
                "required": [
                    "projectId"
                ]
            },
            "ListProjectsResponseContent": {
                "type": "object",
                "properties": {
                    "projects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                }
            },
            "Project": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9]+$"
                    },
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                },
                "required": [
                    "projectId"
                ]
            },
            "UpdateProjectRequestContent": {
                "type": "object",
                "properties": {
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                }
            },
            "UpdateProjectResponseContent": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9]+$"
                    },
                    "client": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "projectLeader": {
                        "type": "string"
                    },
                    "projectManager": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    }
                },
                "required": [
                    "projectId"
                ]
            }
        }
    }
}